Add authorize option
With the vanilla FortuneJS implementation we were unable to verify
whether a consumer had authorization to access/modify data. With
this commit a new 'authorize' option was included to the
Notify Store module.

This option should be a function which returns a promise. Notify
Store will be adding this function to FortuneJS so that it is
invoked with the 'request options' before it (FortuneJS) makes any
requests to the adapter being used.