Expose the Fortune instance used by the store
One of the features of the Store is to provide an additional layer
before the actual request is consumed by Fortune. In this layer a
notify-store consumer can include logic to determine whether the
request is valid, like for example checking for authentication,
authorization, whether the CRUD operation is available, etc... This
can result in the request to get rejected.

When this happens, the consumer can choose to either output a
custom error, which will be represented to the entity which did the
request as an Internal Server Error with his custom message or he
can choose to make use of the Error objects provided by Fortune.

The problem with the latter is that in order for this to work, the
consumer would need a reference of the fortune instance used by
the store. Since Fortune determines what error to send by
reference.

This commit provide the Fortune instance used by the Store.