For each resource the Grant endpoint provided the list of Users who had that particular Grant. As more users get created and assigned Grants, it was going to cause the payload of the Grant resource to be huge. For this reason we decided to remove this info from the Grants endpoint and if a consumer would like to get such info it retrieves it from the users endpoint.
For each resource the States endpoint provided the list of Users who
were in that particular State. As more users get created and assigned
States, it was going to cause the payload of the States resource to
be huge. For this reason we decided to remove this info from the
States endpoint and if a consumer would like to get such info it
retrieves it from the users endpoint.

For each resource the Messages endpoint provided a list of Users who
were marked to not have read the message. Although this info is used
by the client app for when a new message gets created, the payload
of the newly created message is retrieved from Web Socket Server.
This makes the list of users redundant inside the Messages endpoint.

For each resource the Room endpoint provided the list of messages
created within the room. As new messages get created, it was going
to cause the payload of the Room resource to be huge. It was
therefore decided to remove this info from the room resource and
instead use the messages endpoint if a consumer requires the
messages of a particular room.

For each resource the User endpoint provided the list of messages
created by the user. As new messages get created, it was going to
cause the payload of the User resource to be huge. It was
therefore decided to remove this info from the user resource and
isntead use the messages endpoint if a consumer requires the
messages of a particular user.