Make the authorize hook more generic
Currently when using the `authorize` hook, it was only being invoked when the request to FortuneJS was coming from an HTTP Request. This commit changes this so that authorization logic can be done even if the request to FortuneJS is coming from somewhere else.

If the user would like to limit its authorization logic to HTTP Requests, he should include the logic in his custom authorization function using the `options` argument.